+++ packages/core-data/src/selectors.js
@@ -962,3 +962,25 @@ export function __experimentalGetCurrentThemeGlobalStylesVariations( state ) {
 	}
 	return state.themeGlobalStyleVariations[ currentTheme.stylesheet ];
 }
+
+/**
+ * Retrieve the list of registered block patterns.
+ *
+ * @param {Object} state Data state.
+ *
+ * @return {Array} Block pattern list.
+ */
+export function getBlockPatterns( state ) {
+	return state.blockPatterns;
+}
+
+/**
+ * Retrieve the list of registered block pattern categories.
+ *
+ * @param {Object} state Data state.
+ *
+ * @return {Array} Block pattern category list.
+ */
+export function getBlockPatternCategories( state ) {
+	return state.blockPatternCategories;
+}