+++ packages/editor/src/components/provider/use-block-editor-settings.js
@@ -61,6 +61,22 @@ function useBlockEditorSettings( settings, hasTemplate ) {
 		};
 	}, [] );
 
+	const {
+		__experimentalBlockPatterns: settingsBlockPatterns,
+		__experimentalBlockPatternCategories: settingsBlockPatternCategories,
+	} = settings;
+
+	const { blockPatterns, blockPatternCategories } = useSelect(
+		( select ) => ( {
+			blockPatterns:
+				settingsBlockPatterns ?? select( coreStore ).getBlockPatterns(),
+			blockPatternCategories:
+				settingsBlockPatternCategories ??
+				select( coreStore ).getBlockPatternCategories(),
+		} ),
+		[ settingsBlockPatterns, settingsBlockPatternCategories ]
+	);
+
 	const { undo } = useDispatch( editorStore );
 
 	const { saveEntityRecord } = useDispatch( coreStore );
@@ -85,8 +101,6 @@ function useBlockEditorSettings( settings, hasTemplate ) {
 		() => ( {
 			...pick( settings, [
 				'__experimentalBlockDirectory',
-				'__experimentalBlockPatternCategories',
-				'__experimentalBlockPatterns',
 				'__experimentalDiscussionSettings',
 				'__experimentalFeatures',
 				'__experimentalPreferredStyleVariations',
@@ -128,6 +142,8 @@ function useBlockEditorSettings( settings, hasTemplate ) {
 			] ),
 			mediaUpload: hasUploadPermissions ? mediaUpload : undefined,
 			__experimentalReusableBlocks: reusableBlocks,
+			__experimentalBlockPatterns: blockPatterns,
+			__experimentalBlockPatternCategories: blockPatternCategories,
 			__experimentalFetchLinkSuggestions: ( search, searchOptions ) =>
 				fetchLinkSuggestions( search, searchOptions, settings ),
 			__experimentalFetchRichUrlData: fetchUrlData,
@@ -143,6 +159,8 @@ function useBlockEditorSettings( settings, hasTemplate ) {
 			settings,
 			hasUploadPermissions,
 			reusableBlocks,
+			blockPatterns,
+			blockPatternCategories,
 			canUseUnfilteredHTML,
 			undo,
 			hasTemplate,