+++ phpunit/block-supports/typography-test.php
@@ -66,6 +66,80 @@ class WP_Block_Supports_Typography_Test extends WP_UnitTestCase {
 		unregister_block_type( $block_name );
 	}
 
+	function test_typography_with_skipped_serialization_block_supports() {
+		$block_name = 'test/typography-with-skipped-serialization-block-supports';
+		register_block_type(
+			$block_name,
+			array(
+				'api_version' => 2,
+				'attributes'  => array(
+					'style' => array(
+						'type' => 'object',
+					),
+				),
+				'supports'    => array(
+					'typography' => array(
+						'fontSize'                        => true,
+						'lineHeight'                      => true,
+						'__experimentalFontFamily'        => true,
+						'__experimentalLetterSpacing'     => true,
+						'__experimentalSkipSerialization' => true,
+					),
+				),
+			)
+		);
+		$registry   = WP_Block_Type_Registry::get_instance();
+		$block_type = $registry->get_registered( $block_name );
+		$block_atts = array(
+			'style' => array(
+				'typography' => array(
+					'fontSize'      => 'serif',
+					'lineHeight'    => 'serif',
+					'fontFamily'    => '22px',
+					'letterSpacing' => '22px',
+				),
+			),
+		);
+
+		$actual   = gutenberg_apply_typography_support( $block_type, $block_atts );
+		$expected = array();
+
+		$this->assertSame( $expected, $actual );
+		unregister_block_type( $block_name );
+	}
+
+	function test_letter_spacing_with_individual_skipped_serialization_block_supports() {
+		$block_name = 'test/letter-spacing-with-individua-skipped-serialization-block-supports';
+		register_block_type(
+			$block_name,
+			array(
+				'api_version' => 2,
+				'attributes'  => array(
+					'style' => array(
+						'type' => 'object',
+					),
+				),
+				'supports'    => array(
+					'typography' => array(
+						'__experimentalLetterSpacing'     => true,
+						'__experimentalSkipSerialization' => array(
+							'letterSpacing',
+						),
+					),
+				),
+			)
+		);
+		$registry   = WP_Block_Type_Registry::get_instance();
+		$block_type = $registry->get_registered( $block_name );
+		$block_atts = array( 'style' => array( 'typography' => array( 'letterSpacing' => '22px' ) ) );
+
+		$actual   = gutenberg_apply_typography_support( $block_type, $block_atts );
+		$expected = array();
+
+		$this->assertSame( $expected, $actual );
+		unregister_block_type( $block_name );
+	}
+
 	function test_font_family_with_legacy_inline_styles_using_a_css_var() {
 		$block_name = 'test/font-family-with-inline-styles-using-css-var';
 		register_block_type(